Fri, Jul 19, 2024, 8:44 PM -04:00

+++ readability/readability.c
@@ -61,9 +61,9 @@ double get_letter_avg(string text)
     //         printf("NO GO\n");
     //     }
     // }
-    printf("Letter count: %f\n", (num_letters);
-    printf("Word count: %f\n", num_words+1);
-    printf("Sentence count: %f\n", num_sentences);
+    printf("Letter count: %f\n", (float)num_letters);
+    printf("Word count: %f\n", (float)num_words+1);
+    printf("Sentence count: %f\n", (float)num_sentences);
 
     // calculate average # of letters per 100 words
     double letter_average = ((double)num_letters/(double)num_words) * 100;
